+++ src/infra/flow-system-inmem/src/repos/flow/flow_event_store_inmem.rs
@@ -12,7 +12,7 @@ use std::collections::HashMap;
 
 use dill::*;
 use kamu_flow_system::*;
-use opendatafabric::DatasetID;
+use opendatafabric::{AccountName, DatasetID};
 
 use crate::dataset_flow_key::BorrowedFlowKeyDataset;
 
@@ -30,6 +30,8 @@ struct State {
     typed_flows_by_dataset: HashMap<FlowKeyDataset, Vec<FlowID>>,
     all_flows_by_dataset: HashMap<DatasetID, Vec<FlowID>>,
     system_flows_by_type: HashMap<SystemFlowType, Vec<FlowID>>,
+    all_system_flows: Vec<FlowID>,
+    flow_search_index: HashMap<FlowID, FlowIndexEntry>,
     all_flows: Vec<FlowID>,
     last_flow_id: Option<FlowID>,
 }
@@ -45,6 +47,22 @@ impl State {
         self.last_flow_id = Some(next_flow_id);
         next_flow_id
     }
+
+    fn matches_dataset_flow(&self, flow_id: FlowID, filters: &DatasetFlowFilters) -> bool {
+        if let Some(index_entry) = self.flow_search_index.get(&flow_id) {
+            index_entry.matches_dataset_flow_filters(filters)
+        } else {
+            false
+        }
+    }
+
+    fn matches_system_flow(&self, flow_id: FlowID, filters: &SystemFlowFilters) -> bool {
+        if let Some(index_entry) = self.flow_search_index.get(&flow_id) {
+            index_entry.matches_system_flow_filters(filters)
+        } else {
+            false
+        }
+    }
 }
 
 impl EventStoreState<FlowState> for State {
@@ -58,6 +76,79 @@ impl EventStoreState<FlowState> for State {
 
     fn add_event(&mut self, event: FlowEvent) {
         self.events.push(event);
+    }
+}
+
+struct FlowIndexEntry {
+    pub flow_type: AnyFlowType,
+    pub flow_status: FlowStatus,
+    pub initiator: Option<AccountName>,
+}
+
+enum AnyFlowType {
+    Dataset(DatasetFlowType),
+    System(SystemFlowType),
+}
+
+impl FlowIndexEntry {
+    pub fn matches_dataset_flow_filters(&self, filters: &DatasetFlowFilters) -> bool {
+        self.dataset_flow_type_matches(filters.by_flow_type)
+            && self.flow_status_matches(filters.by_flow_status)
+            && self.initiator_matches(filters.by_initiator.as_ref())
+    }
+
+    pub fn matches_system_flow_filters(&self, filters: &SystemFlowFilters) -> bool {
+        self.system_flow_type_matches(filters.by_flow_type)
+            && self.flow_status_matches(filters.by_flow_status)
+            && self.initiator_matches(filters.by_initiator.as_ref())
+    }
+
+    fn dataset_flow_type_matches(
+        &self,
+        maybe_dataset_flow_type_filter: Option<DatasetFlowType>,
+    ) -> bool {
+        match self.flow_type {
+            AnyFlowType::Dataset(dft) => match maybe_dataset_flow_type_filter {
+                Some(flow_type_filter) => flow_type_filter == dft,
+                None => true,
+            },
+            AnyFlowType::System(_) => false,
+        }
+    }
+
+    fn system_flow_type_matches(
+        &self,
+        maybe_system_flow_type_filter: Option<SystemFlowType>,
+    ) -> bool {
+        match self.flow_type {
+            AnyFlowType::Dataset(_) => false,
+            AnyFlowType::System(sft) => match maybe_system_flow_type_filter {
+                Some(flow_type_filter) => flow_type_filter == sft,
+                None => true,
+            },
+        }
+    }
+
+    fn flow_status_matches(&self, maybe_flow_status_filter: Option<FlowStatus>) -> bool {
+        if let Some(flow_status_filter) = maybe_flow_status_filter {
+            flow_status_filter == self.flow_status
+        } else {
+            true
+        }
+    }
+
+    fn initiator_matches(&self, maybe_initiator_filter: Option<&InitiatorFilter>) -> bool {
+        match maybe_initiator_filter {
+            None => true,
+            Some(InitiatorFilter::System) => self.initiator.is_none(),
+            Some(InitiatorFilter::Account(filter_initiator)) => {
+                if let Some(initiator) = &self.initiator {
+                    filter_initiator == initiator
+                } else {
+                    false
+                }
+            }
+        }
     }
 }
 
@@ -91,18 +182,46 @@ impl FlowEventStoreInMem {
                         Entry::Vacant(v) => v.insert(Vec::default()),
                     };
                     all_dataset_entries.push(event.flow_id());
+
+                    state.flow_search_index.insert(
+                        event.flow_id(),
+                        FlowIndexEntry {
+                            flow_type: AnyFlowType::Dataset(flow_key.flow_type),
+                            flow_status: FlowStatus::Waiting,
+                            initiator: e.trigger.initiator_account_name().cloned(),
+                        },
+                    );
                 }
 
                 FlowKey::System(flow_key) => {
-                    let entries = match state.system_flows_by_type.entry(flow_key.flow_type) {
+                    let typed_entries = match state.system_flows_by_type.entry(flow_key.flow_type) {
                         Entry::Occupied(v) => v.into_mut(),
                         Entry::Vacant(v) => v.insert(Vec::default()),
                     };
-                    entries.push(event.flow_id());
+                    typed_entries.push(event.flow_id());
+
+                    state.all_system_flows.push(event.flow_id());
+
+                    state.flow_search_index.insert(
+                        event.flow_id(),
+                        FlowIndexEntry {
+                            flow_type: AnyFlowType::System(flow_key.flow_type),
+                            flow_status: FlowStatus::Waiting,
+                            initiator: e.trigger.initiator_account_name().cloned(),
+                        },
+                    );
                 }
             }
 
             state.all_flows.push(event.flow_id());
+        }
+        /* Existing flow must have been indexed, update status */
+        else if let Some(new_status) = event.new_status() {
+            state
+                .flow_search_index
+                .get_mut(&event.flow_id())
+                .expect("Previously unseen flow ID")
+                .flow_status = new_status;
         }
     }
 }
@@ -149,173 +268,134 @@ impl FlowEventStore for FlowEventStoreInMem {
     }
 
     #[tracing::instrument(level = "debug", skip_all, fields(%dataset_id, ?flow_type))]
-    fn get_last_dataset_flow_of_type(
+    fn get_last_dataset_flow_id_of_type(
         &self,
         dataset_id: &DatasetID,
         flow_type: DatasetFlowType,
-    ) -> Option<FlowID> {
+    ) -> Result<Option<FlowID>, InternalError> {
         let state = self.inner.as_state();
         let g = state.lock().unwrap();
-        g.typed_flows_by_dataset
+        Ok(g.typed_flows_by_dataset
             .get(BorrowedFlowKeyDataset::new(dataset_id, flow_type).as_trait())
-            .and_then(|flows| flows.last().copied())
+            .and_then(|flows| flows.last().copied()))
     }
 
     #[tracing::instrument(level = "debug", skip_all, fields(?flow_type))]
-    fn get_last_system_flow_of_type(&self, flow_type: SystemFlowType) -> Option<FlowID> {
+    fn get_last_system_flow_id_of_type(
+        &self,
+        flow_type: SystemFlowType,
+    ) -> Result<Option<FlowID>, InternalError> {
         let state = self.inner.as_state();
         let g = state.lock().unwrap();
-        g.system_flows_by_type
+        Ok(g.system_flows_by_type
             .get(&flow_type)
-            .and_then(|flows| flows.last().copied())
+            .and_then(|flows| flows.last().copied()))
     }
 
-    #[tracing::instrument(level = "debug", skip_all, fields(%dataset_id, ?flow_type))]
-    fn get_flows_by_dataset_of_type(
+    #[tracing::instrument(level = "debug", skip_all, fields(%dataset_id, ?filters, ?pagination))]
+    fn get_all_flow_ids_by_dataset(
         &self,
         dataset_id: &DatasetID,
-        flow_type: DatasetFlowType,
+        filters: DatasetFlowFilters,
+        pagination: FlowPaginationOpts,
     ) -> FlowIDStream {
-        let dataset_id = dataset_id.clone();
-
-        // TODO: This should be a buffered stream so we don't lock per record
-        Box::pin(async_stream::try_stream! {
-            let borrowed_key = BorrowedFlowKeyDataset::new(&dataset_id, flow_type);
-
-            let mut pos = {
-                let state = self.inner.as_state();
-                let g = state.lock().unwrap();
-                g.typed_flows_by_dataset.get(borrowed_key.as_trait()).map_or(0, Vec::len)
-            };
-
-            loop {
-                if pos == 0 {
-                    break;
-                }
-
-                pos -= 1;
-
-                let next = {
-                    let state = self.inner.as_state();
-                    let g = state.lock().unwrap();
-                    g.typed_flows_by_dataset
-                        .get(borrowed_key.as_trait())
-                        .and_then(|flows| flows.get(pos).copied())
-                };
-
-                let flow_id = match next {
-                    None => break,
-                    Some(flow_id) => flow_id,
-                };
-
-                yield flow_id;
-            }
-        })
-    }
-
-    #[tracing::instrument(level = "debug", skip_all, fields(?flow_type))]
-    fn get_system_flows_of_type(&self, flow_type: SystemFlowType) -> FlowIDStream {
-        let mut pos = {
+        let flow_ids_page: Vec<_> = {
             let state = self.inner.as_state();
             let g = state.lock().unwrap();
-            g.system_flows_by_type.get(&flow_type).map_or(0, Vec::len)
+            g.all_flows_by_dataset
+                .get(dataset_id)
+                .map(|dataset_flow_ids| {
+                    dataset_flow_ids
+                        .iter()
+                        .rev()
+                        .filter(|flow_id| g.matches_dataset_flow(**flow_id, &filters))
+                        .skip(pagination.offset)
+                        .take(pagination.limit)
+                        .map(|flow_id| Ok(*flow_id))
+                        .collect()
+                })
+                .unwrap_or_default()
         };
 
-        // TODO: This should be a buffered stream so we don't lock per record
-        Box::pin(async_stream::try_stream! {
-            loop {
-                if pos == 0 {
-                    break;
-                }
-
-                pos -= 1;
-
-                let next = {
-                    let state = self.inner.as_state();
-                    let g: std::sync::MutexGuard<'_, State> = state.lock().unwrap();
-                    g.system_flows_by_type
-                        .get(&flow_type)
-                        .and_then(|flows| flows.get(pos).copied())
-                };
-
-                let flow_id = match next {
-                    None => break,
-                    Some(flow_id) => flow_id,
-                };
-
-                yield flow_id;
-            }
-        })
+        Box::pin(futures::stream::iter(flow_ids_page))
     }
 
-    #[tracing::instrument(level = "debug", skip_all, fields(%dataset_id))]
-    fn get_all_flows_by_dataset(&self, dataset_id: &DatasetID) -> FlowIDStream {
-        let dataset_id = dataset_id.clone();
+    #[tracing::instrument(level = "debug", skip_all, fields(%dataset_id, ?filters))]
+    async fn get_count_flows_by_dataset(
+        &self,
+        dataset_id: &DatasetID,
+        filters: &DatasetFlowFilters,
+    ) -> Result<usize, InternalError> {
+        let state = self.inner.as_state();
+        let g = state.lock().unwrap();
+        Ok(
+            if let Some(dataset_flow_ids) = g.all_flows_by_dataset.get(dataset_id) {
+                dataset_flow_ids
+                    .iter()
+                    .filter(|flow_id| g.matches_dataset_flow(**flow_id, filters))
+                    .count()
+            } else {
+                0
+            },
+        )
+    }
 
-        // TODO: This should be a buffered stream so we don't lock per record
-        Box::pin(async_stream::try_stream! {
-            let mut pos = {
-                let state = self.inner.as_state();
-                let g = state.lock().unwrap();
-                g.all_flows_by_dataset.get(&dataset_id).map_or(0, Vec::len)
-            };
+    #[tracing::instrument(level = "debug", skip_all, fields(?filters, ?pagination))]
+    fn get_all_system_flow_ids(
+        &self,
+        filters: SystemFlowFilters,
+        pagination: FlowPaginationOpts,
+    ) -> FlowIDStream {
+        let flow_ids_page: Vec<_> = {
+            let state = self.inner.as_state();
+            let g = state.lock().unwrap();
+            g.all_system_flows
+                .iter()
+                .rev()
+                .filter(|flow_id| g.matches_system_flow(**flow_id, &filters))
+                .skip(pagination.offset)
+                .take(pagination.limit)
+                .map(|flow_id| Ok(*flow_id))
+                .collect()
+        };
+        Box::pin(futures::stream::iter(flow_ids_page))
+    }
 
-            loop {
-                if pos == 0 {
-                    break;
-                }
+    #[tracing::instrument(level = "debug", skip_all, fields(?filters))]
+    async fn get_count_system_flows(
+        &self,
+        filters: &SystemFlowFilters,
+    ) -> Result<usize, InternalError> {
+        let state = self.inner.as_state();
+        let g = state.lock().unwrap();
+        Ok(g.all_system_flows
+            .iter()
+            .filter(|flow_id| g.matches_system_flow(**flow_id, filters))
+            .count())
+    }
 
-                pos -= 1;
+    #[tracing::instrument(level = "debug", skip_all, fields(?pagination))]
+    fn get_all_flow_ids(&self, pagination: FlowPaginationOpts) -> FlowIDStream {
+        let flow_ids_page: Vec<_> = {
+            let state = self.inner.as_state();
+            let g = state.lock().unwrap();
+            g.all_flows
+                .iter()
+                .rev()
+                .skip(pagination.offset)
+                .take(pagination.limit)
+                .map(|flow_id| Ok(*flow_id))
+                .collect()
+        };
 
-                let next = {
-                    let state = self.inner.as_state();
-                    let g = state.lock().unwrap();
-                    g.all_flows_by_dataset
-                        .get(&dataset_id)
-                        .and_then(|flows| flows.get(pos).copied())
-                };
-
-                let flow_id = match next {
-                    None => break,
-                    Some(flow_id) => flow_id,
-                };
-
-                yield flow_id;
-            }
-        })
+        Box::pin(futures::stream::iter(flow_ids_page))
     }
 
     #[tracing::instrument(level = "debug", skip_all)]
-    fn get_all_flows(&self) -> FlowIDStream {
-        // TODO: This should be a buffered stream so we don't lock per record
-        Box::pin(async_stream::try_stream! {
-            let mut pos = {
-                let state = self.inner.as_state();
-                let g = state.lock().unwrap();
-                g.all_flows.len()
-            };
-
-            loop {
-                if pos == 0 {
-                    break;
-                }
-
-                pos -= 1;
-
-                let next = {
-                    let state = self.inner.as_state();
-                    let g = state.lock().unwrap();
-                    g.all_flows.get(pos).copied()
-                };
-
-                let flow_id = match next {
-                    None => break,
-                    Some(flow_id) => flow_id,
-                };
-
-                yield flow_id;
-            }
-        })
+    async fn get_count_all_flows(&self) -> Result<usize, InternalError> {
+        let state = self.inner.as_state();
+        let g = state.lock().unwrap();
+        Ok(g.all_flows.len())
     }
 }